refactor: Navigation bar refactor.
