+++ lib/presentation/widgets/widgets.dart
@@ -1,1 +1,2 @@
 export 'items.dart';
+export 'navigationBar.dart';
